requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pandas==2.2.3
numpy<2.0,>=1.24.0
scikit-learn==1.6.0
joblib==1.4.2
python-dotenv==1.0.1

